Implement eventual consistency

+++ src/GymManagement.Application/Subscriptions/Events/SubscriptionDeletedEventHandler.cs
@@ -21,8 +21,13 @@ public class SubscriptionDeletedEventHandler : INotificationHandler<Subscription
 
     public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
     {
-        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId)
-          ?? throw new InvalidOperationException();
+        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId);
+
+        if (subscription is null)
+        {
+            // Resilient error handling
+            throw new InvalidOperationException();
+        }
 
         await _subscriptionsRepository.RemoveSubscriptionAsync(subscription);
